+++ resources/views/livewire/table.blade.php
@@ -108,7 +108,7 @@
             </div>
         </div>
 
-        <x-model-browser::csv-buttons />
+        <x-model-browser::csv-buttons :$exportLimit />
 
     </div>
 </div>